scraper/requirements.txt
httpx==0.28.*
supabase==2.*
python-dotenv==1.*
tenacity==9.*
structlog==24.*
